requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
imageio==2.35.1
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.9.2
networkx==3.3
numpy==2.1.0
opencv-contrib-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
tifffile==2024.8.28
